Editing Path

+++ apps/admin/src/store/selectors/pathSelector.ts
@@ -21,21 +21,32 @@ export const getPathById = selectorFamily({
         if (!(newValue instanceof DefaultValue)) {
             let oldPathSet = new Set(pathState?.path)
             let newPathSet = new Set(newValue?.path)
-            newValue?.path.forEach((pathNode) => {
-                let order = get(getOrder(pathNode))
-                set(getOrder(pathNode), { ...order, driverId: newValue.driverId, driverName: newValue.driverName, currentStatus: "SentToDriver" } as OrderType)
-            })
-            //set difference
-            newPathSet.forEach((value) => {
-                if (oldPathSet.has(value)) {
-                    oldPathSet.delete(value)
-                }
-            })
-            //change status to not assigned
-            oldPathSet.forEach((pathNode) => {
-                let order = get(getOrder(pathNode))
-                set(getOrder(pathNode), { ...order, currentStatus: "NotAssigned" } as OrderType)
-            })
+
+            if (newValue!.driverId != undefined) {
+                newValue?.path.forEach((pathNode) => {
+                    let order = get(getOrder(pathNode))
+                    set(getOrder(pathNode), { ...order, driverId: newValue.driverId, driverName: newValue.driverName, currentStatus: "SentToDriver" } as OrderType)
+                })
+            }
+            if (newValue!.driverId == undefined || newValue!.driverId == null) {
+                newValue?.path.forEach((node) => {
+                    let order = get(getOrder(node))
+                    set(getOrder(node), { ...order, currentStatus: "PathAssigned", assignedPathId: newValue.pathId } as OrderType)
+                })
+                //set difference
+                newPathSet.forEach((value) => {
+                    if (oldPathSet.has(value)) {
+                        oldPathSet.delete(value)
+                    }
+                })
+                //change status to not assigned
+                oldPathSet.forEach((pathNode) => {
+                    let order = get(getOrder(pathNode))
+                    set(getOrder(pathNode), { ...order, currentStatus: "NotAssigned", assignedPathId: "" } as OrderType)
+                })
+            }
+
+
 
         }
